Requirements Refactored
Newtonsoft changed System.text.json
history models created and moved
Relation deleted for history tables
