+++ amorphie.contract.application/Contract/ContractMapProfile.cs
@@ -1,5 +1,6 @@
 ﻿using amorphie.contract.application.Contract.Dto;
 using amorphie.contract.core.Entity.Contract;
+using amorphie.contract.core.Model.History;
 using AutoMapper;
 
 namespace amorphie.contract.application.Contract
@@ -30,6 +31,6 @@ namespace amorphie.contract.application.Contract
                     .ReverseMap();
 
         }
-    }
+    }  
 
 }
