+++ amorphie.contract.core/Entity/Document/DocumentGroups/DocumentGroup.cs
@@ -24,7 +24,6 @@ namespace amorphie.contract.core.Entity.Document.DocumentGroups
         public EStatus Status { get; set; } = default!;
         public ICollection<DocumentGroupDetail> DocumentGroupDetails { get; set; } = new List<DocumentGroupDetail>();
         public ICollection<DocumentGroupLanguageDetail> DocumentGroupLanguageDetail { get; set; } = new List<DocumentGroupLanguageDetail>();
-        public ICollection<DocumentGroupHistory>? DocumentGroupHistories { get; set; } = new List<DocumentGroupHistory>();
 
     }
 }
