+++ amorphie.contract.core/Entity/Document/DocumentGroups/DocumentGroupHistory.cs
@@ -6,6 +6,7 @@ using System.ComponentModel.DataAnnotations.Schema;
 using System.Text.Json;
 using amorphie.contract.core.Entity.Document.DocumentGroups;
 using amorphie.contract.core.Enum;
+using amorphie.contract.core.Model.History;
 
 namespace amorphie.contract.core.Entity.Contract
 {
@@ -15,17 +16,5 @@ namespace amorphie.contract.core.Entity.Contract
         [Required]
         public DocumentGroupHistoryModel DocumentGroupHistoryModel { get; set; }
         public Guid DocumentGroupId { get; set; }
-        [ForeignKey("DocumentGroupId")]
-
-        public DocumentGroup DocumentGroup { get; set; } = default!;
-    }
-
-    public class DocumentGroupHistoryModel : BaseEntity
-    {
-        public EStatus Status { get; set; } = default!;
-        public ICollection<DocumentGroupDetail> DocumentGroupDetails { get; set; } = new List<DocumentGroupDetail>();
-        public ICollection<DocumentGroupLanguageDetail> DocumentGroupLanguageDetail { get; set; } = new List<DocumentGroupLanguageDetail>();
-        public ICollection<DocumentGroupHistory>? DocumentGroupHistories { get; set; } = new List<DocumentGroupHistory>();
-
     }
 }
