+++ amorphie.contract.infrastructure/Configurations/Document/DocumentGroups/DocumentGroupHistoryConfiguration.cs
@@ -1,13 +1,9 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
 using System.Text.Json;
 using System.Text.Json.Serialization;
-using System.Threading.Tasks;
 using amorphie.contract.core.Entity.Contract;
+using amorphie.contract.core.Model.History;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using Newtonsoft.Json;
 namespace amorphie.contract.infrastructure.Configurations.Contract
 {
     public class DocumentGroupHistoryConfiguration : ConfigurationBaseAudiEntity<DocumentGroupHistory>,
@@ -15,17 +11,17 @@ namespace amorphie.contract.infrastructure.Configurations.Contract
     {
         public virtual void Configure(EntityTypeBuilder<DocumentGroupHistory> builder)
         {
-            var settings = new JsonSerializerSettings
-            {
-                ReferenceLoopHandling = ReferenceLoopHandling.Ignore
-            };
+            var options = new JsonSerializerOptions
+                {
+                    ReferenceHandler = ReferenceHandler.IgnoreCycles
+                };
 
-            builder.Property(e => e.DocumentGroupHistoryModel)
-                .HasColumnType("jsonb")
-                .HasConversion(
-                    v => JsonConvert.SerializeObject(v, settings),
-                    v => JsonConvert.DeserializeObject<DocumentGroupHistoryModel>(v) 
-                );
+                builder.Property(e => e.DocumentGroupHistoryModel)
+                    .HasColumnType("jsonb")
+                    .HasConversion(
+                        v => JsonSerializer.Serialize(v, options),
+                        v => JsonSerializer.Deserialize<DocumentGroupHistoryModel>(v, options)
+                    );
         }
     }
 }
